experiments/media_summarizer/pyproject.toml
[project]
name = "media-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit==1.32.2",
    "google-generativeai==0.3.2",
    "python-dotenv==1.0.1",
    "moviepy==1.0.3",
    "speechrecognition==3.10.1",
    "pydub==0.25.1",
    "google-genai>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
